+++ pkg/ddl/table_mode_test.go
@@ -21,14 +21,12 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/pkg/ddl"
+	"github.com/pingcap/tidb/pkg/ddl/testutil"
 	"github.com/pingcap/tidb/pkg/domain"
 	"github.com/pingcap/tidb/pkg/errno"
-	"github.com/pingcap/tidb/pkg/kv"
-	"github.com/pingcap/tidb/pkg/meta"
 	"github.com/pingcap/tidb/pkg/meta/model"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/terror"
-	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/stretchr/testify/require"
 )
@@ -42,42 +40,6 @@ func getClonedTableInfoFromDomain(
 	tbl, err := dom.InfoSchema().TableByName(context.Background(), ast.NewCIStr(dbName), ast.NewCIStr(tableName))
 	require.NoError(t, err)
 	return tbl.Meta().Clone()
-}
-
-func setTableModeTest(
-	ctx sessionctx.Context,
-	t *testing.T,
-	store kv.Storage,
-	de ddl.Executor,
-	dbInfo *model.DBInfo,
-	tblInfo *model.TableInfo,
-	mode model.TableMode,
-) error {
-	args := &model.AlterTableModeArgs{
-		TableMode: mode,
-		SchemaID:  dbInfo.ID,
-		TableID:   tblInfo.ID,
-	}
-	err := de.AlterTableMode(ctx, args)
-	if err == nil {
-		testCheckTableState(t, store, dbInfo, tblInfo, model.StatePublic)
-		checkTableModeTest(t, store, dbInfo, tblInfo, mode)
-	}
-
-	return err
-}
-
-func checkTableModeTest(t *testing.T, store kv.Storage, dbInfo *model.DBInfo, tblInfo *model.TableInfo, mode model.TableMode) {
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnDDL)
-	err := kv.RunInNewTxn(ctx, store, false, func(ctx context.Context, txn kv.Transaction) error {
-		tt := meta.NewMutator(txn)
-		info, err := tt.GetTable(dbInfo.ID, tblInfo.ID)
-		require.NoError(t, err)
-		require.NotNil(t, info)
-		require.Equal(t, mode, info.Mode)
-		return nil
-	})
-	require.NoError(t, err)
 }
 
 func checkErrorCode(t *testing.T, err error, expected int) {
@@ -109,7 +71,7 @@ func TestTableModeBasic(t *testing.T) {
 	require.NoError(t, err)
 	dbInfo, ok := domain.InfoSchema().SchemaByName(ast.NewCIStr("test"))
 	require.True(t, ok)
-	checkTableModeTest(t, store, dbInfo, tblInfo, model.TableModeImport)
+	testutil.TestCheckTableMode(t, store, dbInfo, tblInfo, model.TableModeImport)
 	// not allow delete foreign key constraint
 	tk.MustGetErrCode("ALTER TABLE t1_foreign_key DROP FOREIGN KEY fk_1", errno.ErrProtectedTableMode)
 
@@ -121,7 +83,7 @@ func TestTableModeBasic(t *testing.T) {
 	require.NoError(t, err)
 	dbInfo, ok = domain.InfoSchema().SchemaByName(ast.NewCIStr("test"))
 	require.True(t, ok)
-	checkTableModeTest(t, store, dbInfo, tblInfo, model.TableModeRestore)
+	testutil.TestCheckTableMode(t, store, dbInfo, tblInfo, model.TableModeRestore)
 
 	// For testing accessing table metadata is allowed when table is in ModeRestore
 	tk.MustExec("show create table t1_restore_import")
@@ -168,24 +130,24 @@ func TestTableModeBasic(t *testing.T) {
 	tk.MustExec("rollback")
 
 	// For testing AlterTable ModeRestore -> ModeImport is not allowed
-	err = setTableModeTest(ctx, t, store, de, dbInfo, tblInfo, model.TableModeImport)
+	err = testutil.TestSetTableMode(ctx, t, store, de, dbInfo, tblInfo, model.TableModeImport)
 	require.ErrorContains(t, err, "Invalid mode set from (or by default) Restore to Import for table t1_restore_import")
 
 	// For testing AlterTableMode ModeRestore -> ModeNormal
-	err = setTableModeTest(ctx, t, store, de, dbInfo, tblInfo, model.TableModeNormal)
+	err = testutil.TestSetTableMode(ctx, t, store, de, dbInfo, tblInfo, model.TableModeNormal)
 	require.NoError(t, err)
 
 	// For testing AlterTableMode ModeNormal -> ModeRestore
-	err = setTableModeTest(ctx, t, store, de, dbInfo, tblInfo, model.TableModeRestore)
+	err = testutil.TestSetTableMode(ctx, t, store, de, dbInfo, tblInfo, model.TableModeRestore)
 	require.NoError(t, err)
 	// For testing AlterTableMode ModeRestore -> ModeRestore
-	err = setTableModeTest(ctx, t, store, de, dbInfo, tblInfo, model.TableModeRestore)
+	err = testutil.TestSetTableMode(ctx, t, store, de, dbInfo, tblInfo, model.TableModeRestore)
 	require.NoError(t, err)
 
 	// For testing an exist table with ModeImport is not allowed recreate with ModeRestore from BR
-	err = setTableModeTest(ctx, t, store, de, dbInfo, tblInfo, model.TableModeNormal)
+	err = testutil.TestSetTableMode(ctx, t, store, de, dbInfo, tblInfo, model.TableModeNormal)
 	require.NoError(t, err)
-	err = setTableModeTest(ctx, t, store, de, dbInfo, tblInfo, model.TableModeImport)
+	err = testutil.TestSetTableMode(ctx, t, store, de, dbInfo, tblInfo, model.TableModeImport)
 	require.NoError(t, err)
 	tblInfo.Mode = model.TableModeRestore
 	err = de.CreateTableWithInfo(tk.Session(), ast.NewCIStr("test"), tblInfo, nil, ddl.WithOnExist(ddl.OnExistIgnore))
@@ -209,9 +171,9 @@ func TestTableModeBasic(t *testing.T) {
 		ddl.WithOnExist(ddl.OnExistIgnore),
 	)
 	require.NoError(t, err)
-	checkTableModeTest(t, store, dbInfo, tblInfo1, model.TableModeNormal)
-	checkTableModeTest(t, store, dbInfo, tblInfo2, model.TableModeImport)
-	checkTableModeTest(t, store, dbInfo, tblInfo3, model.TableModeRestore)
+	testutil.TestCheckTableMode(t, store, dbInfo, tblInfo1, model.TableModeNormal)
+	testutil.TestCheckTableMode(t, store, dbInfo, tblInfo2, model.TableModeImport)
+	testutil.TestCheckTableMode(t, store, dbInfo, tblInfo3, model.TableModeRestore)
 }
 
 func TestTableModeConcurrent(t *testing.T) {
@@ -236,7 +198,7 @@ func TestTableModeConcurrent(t *testing.T) {
 	for _, info := range t1Infos {
 		go func(info *model.TableInfo) {
 			defer wg.Done()
-			errs <- setTableModeTest(ctx, t, store, de, dbInfo, info, model.TableModeImport)
+			errs <- testutil.TestSetTableMode(ctx, t, store, de, dbInfo, info, model.TableModeImport)
 		}(info)
 	}
 	wg.Wait()
@@ -252,7 +214,7 @@ func TestTableModeConcurrent(t *testing.T) {
 	}
 	require.Equal(t, 2, successCount)
 	require.Nil(t, failedErr)
-	checkTableModeTest(t, store, dbInfo, t1Infos[0], model.TableModeImport)
+	testutil.TestCheckTableMode(t, store, dbInfo, t1Infos[0], model.TableModeImport)
 
 	// Concurrency test2: concurrently alter t1 to ModeNormal, expecting both success.
 	t1NormalInfos := []*model.TableInfo{
@@ -266,7 +228,7 @@ func TestTableModeConcurrent(t *testing.T) {
 		go func(info *model.TableInfo) {
 			defer wg2.Done()
 
-			errs2 <- setTableModeTest(ctx, t, store, de, dbInfo, info, model.TableModeNormal)
+			errs2 <- testutil.TestSetTableMode(ctx, t, store, de, dbInfo, info, model.TableModeNormal)
 		}(info)
 	}
 	wg2.Wait()
@@ -274,7 +236,7 @@ func TestTableModeConcurrent(t *testing.T) {
 	for e := range errs2 {
 		require.NoError(t, e)
 	}
-	checkTableModeTest(t, store, dbInfo, t1NormalInfos[0], model.TableModeNormal)
+	testutil.TestCheckTableMode(t, store, dbInfo, t1NormalInfos[0], model.TableModeNormal)
 
 	// Concurrency test3: concurrently alter t1 to ModeRestore, expecting both success.
 	t1Infos = []*model.TableInfo{
@@ -287,7 +249,7 @@ func TestTableModeConcurrent(t *testing.T) {
 	for _, info := range t1Infos {
 		go func(info *model.TableInfo) {
 			defer wg3.Done()
-			errs <- setTableModeTest(ctx, t, store, de, dbInfo, info, model.TableModeRestore)
+			errs <- testutil.TestSetTableMode(ctx, t, store, de, dbInfo, info, model.TableModeRestore)
 		}(info)
 	}
 	wg3.Wait()
@@ -303,7 +265,7 @@ func TestTableModeConcurrent(t *testing.T) {
 	}
 	require.Equal(t, 2, successCount)
 	require.Nil(t, failedErr)
-	checkTableModeTest(t, store, dbInfo, t1Infos[0], model.TableModeRestore)
+	testutil.TestCheckTableMode(t, store, dbInfo, t1Infos[0], model.TableModeRestore)
 
 	// Concurrency test4: concurrently alter t1 to ModeRestore and ModeImport, expecting one success, one failure.
 	modes := []model.TableMode{
@@ -320,7 +282,7 @@ func TestTableModeConcurrent(t *testing.T) {
 	for i, mode := range modes {
 		go func(clone *model.TableInfo, m model.TableMode) {
 			defer wg4.Done()
-			errs <- setTableModeTest(ctx, t, store, de, dbInfo, clone, m)
+			errs <- testutil.TestSetTableMode(ctx, t, store, de, dbInfo, clone, m)
 		}(clones[i], mode)
 	}
 	wg4.Wait()
