+++ pkg/ddl/tests/fk/BUILD.bazel
@@ -12,16 +12,15 @@ go_test(
     deps = [
         "//pkg/config",
         "//pkg/ddl",
+        "//pkg/ddl/testutil",
         "//pkg/domain",
         "//pkg/errno",
         "//pkg/infoschema",
-        "//pkg/kv",
         "//pkg/meta",
         "//pkg/meta/autoid",
         "//pkg/meta/model",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
-        "//pkg/sessionctx",
         "//pkg/sessiontxn",
         "//pkg/testkit",
         "//pkg/testkit/testsetup",
