+++ pkg/ddl/testutil/BUILD.bazel
@@ -9,14 +9,17 @@ go_library(
     importpath = "github.com/pingcap/tidb/pkg/ddl/testutil",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/ddl",
         "//pkg/ddl/logutil",
         "//pkg/disttask/operator",
         "//pkg/domain",
         "//pkg/kv",
+        "//pkg/meta",
         "//pkg/meta/model",
         "//pkg/parser/ast",
         "//pkg/session",
         "//pkg/session/types",
+        "//pkg/sessionctx",
         "//pkg/sessiontxn",
         "//pkg/table",
         "//pkg/table/tables",
